+++ rgbToHex.js
@@ -1,0 +1,11 @@
+console.log("rgb to hex loaded");
+
+//for RGB to Hex
+function componentToHex(c) {
+  var hex = c.toString(16);
+  return hex.length == 1 ? "0" + hex : hex;
+}
+
+module.exports.rgbToHex = function(r, g, b) {
+  return "#" + componentToHex(r) + componentToHex(g) + componentToHex(b);
+}
